Fix regression causing scheduling rate limiter to not be honored.
Also added findbugs to the build.

Bugs closed: AURORA-416

Reviewed at https://reviews.apache.org/r/21352/

+++ src/test/java/org/apache/aurora/scheduler/thrift/SchedulerThriftInterfaceTest.java
@@ -957,7 +957,7 @@ public class SchedulerThriftInterfaceTest extends EasyMockTest {
 
   @Test
   public void testGetJobSummary() throws Exception {
-    int nextCronRunMs = 100;
+    long nextCronRunMs = 100;
     TaskConfig ownedCronJobTask = nonProductionTask()
         .setJobName(JobKeys.TO_JOB_NAME.apply(JOB_KEY))
         .setOwner(ROLE_IDENTITY)
